+++ Source/MindKeeper.Shared/Models/ApiModels/Statistics/StatsUserResult.cs
@@ -5,7 +5,7 @@ namespace MindKeeper.Shared.Models.ApiModels.Statistics
     public class StatsUserResult
     {
         public Guid UserId { get; set; }
-        public int IdeasCreatedCount { get; set; }
-        public int AchievementsCount { get; set; }
+        public long IdeasCreatedCount { get; set; }
+        public long AchievementsCount { get; set; }
     }
 }
